+++ api/__init__.py
@@ -2,3 +2,9 @@
 """
 API 模塊，負責與交易所API的通訊
 """
+from .base_client import BaseExchangeClient
+from .bp_client import BPClient
+from .aster_client import AsterClient
+from .paradex_client import ParadexClient
+
+__all__ = ['BaseExchangeClient', 'BPClient', 'AsterClient', 'ParadexClient']